requirements.txt
BigchainDB==2.0.0a6
bigchaindb-abci==0.4.5
bigchaindb-driver==0.5.0a1
bitarray==0.8.1
blaze==0.10.1
bokeh==0.12.4
boto==2.45.0
Bottleneck==1.2.0
certifi==2018.4.16
cffi==1.9.1
chardet==3.0.4
chest==0.2.3
click==6.7
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.7
colorlog==3.1.4
conda==4.3.14
configobj==5.0.6
contextlib2==0.5.4
coverage==4.5.1
cryptoconditions==0.6.0.dev1
cryptography==1.9
cycler==0.10.0
Cython==0.25.2
cytoolz==0.9.0.1
dask==0.13.0
datashape==0.5.4
decorator==4.0.11
dill==0.2.5
docutils==0.13.1
et-xmlfile==1.0.1
eth-abi==1.1.1
eth-account==0.2.2
eth-hash==0.1.3
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-utils==1.0.3
fastcache==1.0.2
Flask==0.12
Flask-Cors==3.0.2
Flask-RESTful==0.3.6
gevent==1.3.1
greenlet==0.4.13
gunicorn==19.8.1
h5py==2.6.0
HeapDict==1.0.0
hexbytes==0.1.0
idna==2.6
idna-ssl==1.0.1
imagesize==0.7.1
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.9.0
Jinja2==2.9.4
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
lazy-object-proxy==1.2.2
llvmlite==0.15.0
locket==0.2.0
logstats==0.2.1
lru-dict==1.1.6
lxml==3.7.2
MarkupSafe==0.23
matplotlib==2.0.0
mistune==0.7.3
more-itertools==4.1.0
mpmath==0.19
multidict==4.3.1
multipipes==0.1.0
multipledispatch==0.4.9
nbconvert==4.2.0
nbformat==4.2.0
networkx==1.11
nltk==3.2.2
nose==1.3.7
notebook==4.3.1
numba==0.30.1
numexpr==2.6.1
numpy==1.11.3
numpydoc==0.6.0
odo==0.5.0
openpyxl==2.4.1
pandas==0.19.2
parsimonious==0.8.0
partd==0.3.7
pathlib2==2.2.0
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
pluggy==0.6.0
ply==3.9
prompt-toolkit==1.0.9
protobuf==3.5.2.post1
psutil==5.0.1
ptyprocess==0.5.1
py==1.5.3
pyasn1==0.2.3
pycosat==0.6.1
pycparser==2.17
pycrypto==2.6.1
pycryptodome==3.6.1
pycurl==7.43.0
pyflakes==1.5.0
Pygments==2.1.3
pylint==1.6.4
pymongo==3.6.1
PyNaCl==1.1.2
pyOpenSSL==16.2.0
pyparsing==2.1.4
pysha3==1.0.2
pytest==3.6.0
pytest-cov==2.5.1
pytest-pythonpath==0.7.2
python-dateutil==2.6.0
python-rapidjson==0.5.2
python-rapidjson-schema==0.1.1
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.3
qtconsole==4.2.1
QtPy==1.2.1
redis==2.10.5
requests==2.18.4
rethinkdb==2.3.0.post6
rlp==1.0.1
rope-py3k==0.9.4.post1
scikit-image==0.12.3
scikit-learn==0.18.1
scipy==0.18.1
seaborn==0.7.1
setproctitle==1.1.10
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
Sphinx==1.5.1
spyder==3.1.2
SQLAlchemy==1.1.5
statsd==3.2.1
statsmodels==0.6.1
sympy==1.0
tables==3.3.0
terminado==0.6
toolz==0.8.2
tornado==4.4.2
traitlets==4.3.1
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
web3==4.2.1
websockets==5.0
Werkzeug==0.11.15
widgetsnbextension==1.2.6
wrapt==1.10.8
xlrd==1.0.0
XlsxWriter==0.9.6
xlwt==1.2.0
yarl==1.2.4
